+++ web-app/src/App.js
@@ -3,7 +3,7 @@ import Navbar from './Components/Navbar/index';
 import Home from './Components/Home';
 import Search from './Components/Search';
 import SightDetails from './Components/Sights/Details';
-import SightForm from './Components/SightForm';
+import SightForm from './Components/Sights/New/Form';
 import Login from './Components/Login';
 import Register from './Components/Register';
 import Administration from './Components/Administration';
@@ -17,7 +17,7 @@ const App = () => (
       <Route exact path="/" component={Home} />
       <Route path="/home" component={Home} />
       <Route path="/sights" component={Search} />
-      <Route path="/add-sight" component={SightForm} />
+      <Route path="/sight/add" component={SightForm} />
       <Route path="/details/:id" component={SightDetails} />
       <Route path="/login" component={Login} />
       <Route path="/register" component={Register} />
